recommender/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-background-tasks==1.2.0
django-compat==1.0.15
django-cors-headers==2.4.0
django-rest==0.0.1
djangorestframework==3.8.2
idna==2.7
iteration-utilities==0.7.0
joblib==0.13.2
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
python-dateutil==2.8.0
pytz==2018.9
requests==2.20.1
requests-oauthlib==1.0.0
scipy==1.2.1
six==1.12.0
ticketpy==1.1.2
urllib3==1.24.1
